+++ core/application.go
@@ -36,6 +36,8 @@ type Application struct {
 	authman       Authman
 	corebb        Core
 	rewards       Rewards
+
+	authmanSyncInProgress bool
 }
 
 // Start starts the corebb part of the application